cleanup: Move class namespace
