+++ src/ExtractFeedlyOPML.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Text;
 using System.Threading.Tasks;
+using FeedlyOpmlExport.Functions.Utilities;
 using Microsoft.Azure.WebJobs;
 using Microsoft.Extensions.Logging;
