+++ src/RefreshFeedlyAuthToken.cs
@@ -2,6 +2,7 @@ using System;
 using Microsoft.Azure.WebJobs;
 using Microsoft.Extensions.Logging;
 using System.Threading.Tasks;
+using FeedlyOpmlExport.Functions.Utilities;
 
 namespace FeedlyOpmlExport.Functions
 {
